Relocate run parameters test to models and its fixtures to the common conftest
