+++ tests/models/demultiplexing/test_run_parameters.py
@@ -18,7 +18,7 @@ def test_get_flowcell_type_hiseq_run_parameters(hiseq_run_parameters: Path):
 
 
 def test_get_flowcell_type_novaseq_run_parameters(novaseq_run_parameters: Path):
-    # GIVEN the path to a file with hiseq run parameters
+    # GIVEN the path to a file with novaseq run parameters
     # GIVEN a RunParameters object
     run_parameters_object = RunParameters(novaseq_run_parameters)
